pyproject.toml
[tool.poetry]
name = "autocapture"
version = "0.1.0"
description = "High-performance local activity capture and recall pipeline"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "autocapture", exclude = ["autocapture/_legacy/**"] }]
include = ["autocapture/prompts/**/*.yaml", "autocapture/ui/web/**"]

[tool.poetry.scripts]
autocapture = "autocapture.main:main"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
pyyaml = "^6.0.1"
loguru = "^0.7.2"
pydantic = "^2.6.4"
fastapi = "^0.115.0"
uvicorn = {version = "^0.30.0", extras = ["standard"]}
httpx = "^0.27.0"
pyside6 = "^6.7.0"
pynput = "^1.7.6"
sqlalchemy = "^2.0.29"
psycopg = { version = "^3.1.18", extras = ["binary"] }
pillow = "^10.2.0"
numpy = "^1.26.4"
opencv-python-headless = "^4.9.0.80"
qdrant-client = "^1.8.2"
sentence-transformers = "^2.4.0"
fastembed = "^0.3.3"
hnswlib = "^0.8.0"
uvloop = {version = "^0.19.0", markers = "sys_platform != 'win32'"}
prometheus-client = "^0.20.0"
cryptography = "^42.0.5"
psutil = "^5.9.8"
dxcam = "^0.0.5"
mss = "^9.0.1"
rapidocr-onnxruntime = "^1.3.24"
openai = "^1.14.3"
python-dateutil = "^2.9.0"
PyJWT = "^2.8.0"
alembic = "^1.13.1"

[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
ruff = "^0.3.4"
pyinstaller = "^6.6.0"
pytest = "^8.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
